Rename the TimeLineComponent to TimelineComponent
